+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Draw a separator between albums with the same name, but a different artist.
 * Suppress output of all external commands.
 * Consider mouse support when pausing and unpausing curses interface.
+* Reduce CPU usage of the frequency spectrum visualizer.
 
 # ncmpcpp-0.9 (2020-12-20)
 * Fix various Mopidy specific bugs.